streamlit_app/requirements.txt
streamlit==1.53.1
google-generativeai==0.8.6
requests==2.32.5
tenacity==9.1.2
google-cloud-secret-manager==2.22.0
